requirements.txt
pybadges==3.0.1
Pillow==10.1.0
CairoSVG==2.7.1
regex==2023.3.23
rich==13.7.0
lingua-language-detector==2.0.2

